[Back] Users (feat): add checks when creating/deleting user or list.

+++ src/Backend/DataManagement/Users/Services/UsersManagementService.cs
@@ -3,7 +3,7 @@ using Microsoft.EntityFrameworkCore;
 
 namespace Backend.DataManagement.Users.Services;
 
-public class UsersManagementService(UsersContext context)
+public class UsersManagementService(UsersContext context, ILogger<UsersManagementService> logger)
 {
     public async ValueTask<User?> FindByIdAsync(
         Guid              id,
@@ -19,14 +19,12 @@ public class UsersManagementService(UsersContext context)
 
         if (includeAnalyticsLists)
         {
-            user.AnalyticsLists = context.AnalyticsLists.Where(list => list.CreatorId == id)
-                                         .AsEnumerable();
+            user.AnalyticsLists = context.AnalyticsLists.Where(list => list.CreatorId == id).ToList();
         }
 
         if (includeTeams)
         {
-            user.OrganisedTeams = context.Teams.Where(team => team.OrganiserId == id)
-                                         .AsEnumerable();
+            user.OrganisedTeams = context.Teams.Where(team => team.OrganiserId == id).ToList();
         }
 
         return user;
@@ -46,21 +44,26 @@ public class UsersManagementService(UsersContext context)
 
         if (includeAnalyticsLists)
         {
-            user.AnalyticsLists = context.AnalyticsLists.Where(list => list.CreatorId == user.Id)
-                                         .AsEnumerable();
+            user.AnalyticsLists = context.AnalyticsLists.Where(list => list.CreatorId == user.Id).ToList();
         }
 
         if (includeTeams)
         {
-            user.OrganisedTeams = context.Teams.Where(team => team.OrganiserId == user.Id)
-                                         .AsEnumerable();
+            user.OrganisedTeams = context.Teams.Where(team => team.OrganiserId == user.Id).ToList();
         }
 
         return user;
     }
 
-    public async Task<User> CreateUserAsync(Guid id, string name, CancellationToken cancellationToken)
+    public async Task<User?> TryCreateUserAsync(Guid id, string name, CancellationToken cancellationToken)
     {
+        User? found = await context.Users.SingleOrDefaultAsync(u => u.Name == name, cancellationToken);
+        if (found is not null)
+        {
+            logger.LogDebug("User with name: {Name} already registered", name);
+            return null;
+        }
+        
         User user = new(id, name);
 
         await context.Users.AddAsync(user, cancellationToken);
@@ -69,10 +72,15 @@ public class UsersManagementService(UsersContext context)
         return user;
     }
 
-    public async Task<bool> DeleteUserAsync(Guid id, CancellationToken cancellationToken)
+    public async Task<bool> TryDeleteUserAsync(Guid id, CancellationToken cancellationToken)
     {
-        User user = await context.Users.FindAsync([ id ], cancellationToken)
-                 ?? throw new InvalidOperationException($"User with ID: {id} can't be found");
+        User? user = await context.Users.FindAsync([ id ], cancellationToken);
+        if (user is null)
+        {
+            logger.LogDebug("User with id: {ID} can't be found for delete", id);
+            return false;
+        }
+        
         context.Users.Remove(user);
 
         int rowsDeleted = await context.SaveChangesAsync(cancellationToken);
